refactor: Implement 'Lane' class
The 'Lane' class stores all information about the left and right lines. 
The main requirement to implement the class, was to store history line data to allow line-smoothing. 
The method 'line_search' performs the the 'window' and 'ancestor' line search.
The method 'get_radius_annotations' calculates the line curvature radiuses and returns them in a formated way.